Module bikeshare::types

Structs

Api
Both endpoints used by this application has the same top level layout, this struct generalized over them both.

ApiData
Both endpoints used by this application has the same data layout with the stations key, this struct generalized over them both.

Information
The data inside the station_information endpoint is modelled by this struct

JoinedStation
The status struct contains "dynamic" content about stations and the information endpoint has the "static" content. This struct represents a join of them both. The status content may be missing.

JoinedStatus
The status content of JoinedStation.

Status
The data inside the station_status endpoint is modelled by this struct

Constants

_IMPL_DESERIALIZE_FOR_Api

_IMPL_DESERIALIZE_FOR_ApiData

_IMPL_DESERIALIZE_FOR_Information

_IMPL_DESERIALIZE_FOR_JoinedStation

_IMPL_DESERIALIZE_FOR_JoinedStatus

_IMPL_DESERIALIZE_FOR_Status

_IMPL_SERIALIZE_FOR_Api

_IMPL_SERIALIZE_FOR_ApiData

_IMPL_SERIALIZE_FOR_Information

_IMPL_SERIALIZE_FOR_JoinedStation

_IMPL_SERIALIZE_FOR_JoinedStatus

_IMPL_SERIALIZE_FOR_Status